requirements.txt
# Usage: pip install -r requirements.txt
matplotlib>=3.3
onnxruntime
exif
numpy>=1.18.5
opencv-python>=4.1.1
Pillow==9.5.0
onnx==1.14.0
onnxruntime==1.15.1
psutil  # system resources
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
torch>=1.7.0  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.8.1
tqdm>=4.64.0
ultralytics>=8.0.111
geopy
folium
pandas
streamlit
streamlit_option_menu
plotly
streamlit-folium
fpdf



